+++ rmgpy/data/kinetics/groups.py
@@ -426,7 +426,7 @@ class KineticsGroups(Database):
 
                 # Determine error in each group (on log scale)
                 stdev = np.zeros(len(group_list) + 1, float)
-                count = np.zeros(len(group_list) + 1, np.int)
+                count = np.zeros(len(group_list) + 1, int)
 
                 for index in range(len(training_set)):
                     template, kinetics = training_set[index]
